enable peephole opt on windows, make sure AVMPLUS_DISABLE_MIR works

+++ core/Translator.cpp
@@ -1018,7 +1018,7 @@ namespace avmplus
 				// Backward branch
 				AvmAssert((int32)I[nextI - 1][1] < 0);
 				// Install the negative of the absolute word offset of the target
-				I[nextI - 1][1] = -(buffer_offset + (dest - buffers->data) + (int32)I[nextI - 1][1]);
+				I[nextI - 1][1] = -int32(buffer_offset + (dest - buffers->data) + (int32)I[nextI - 1][1]);
 			}
 		}
 		
@@ -1067,7 +1067,7 @@ namespace avmplus
 					// Backward jump
 					// Compute new jump offset
 					AvmAssert(!forward);
-					*dest = -(buffer_offset + (dest + 1 - buffers->data) + offset);
+					*dest = -int32(buffer_offset + (dest + 1 - buffers->data) + offset);
 					dest++;
 				}
 			}
